+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -37,6 +37,7 @@ namespace UnityEngine.Rendering.HighDefinition
         SSGITrace,
         SSGIDenoise,
         SSGIUpscale,
+        SSGIConvert,
 
         ForwardEmissive,
         ForwardOpaque,
@@ -73,6 +74,7 @@ namespace UnityEngine.Rendering.HighDefinition
         DisplayPlanarReflectionProbeAtlas,
         BlitTextureInPotAtlas,
         AreaLightCookieConvolution,
+        DisplayDensityVolumeAtlas,
 
         UpdateSkyEnvironmentConvolution,
         RenderSkyToCubemap,
@@ -80,6 +82,7 @@ namespace UnityEngine.Rendering.HighDefinition
         UpdateSkyAmbientProbe,
         PreRenderSky,
         RenderSky,
+        RenderClouds,
         OpaqueAtmosphericScattering,
         InScatteredRadiancePrecomputation,
 
@@ -87,6 +90,7 @@ namespace UnityEngine.Rendering.HighDefinition
         VolumetricLighting,
         VolumetricLightingFiltering,
         PrepareVisibleDensityVolumeList,
+        UpdateDensityVolumeAtlas,
 
         // RT Cluster
         RaytracingBuildCluster,
@@ -95,8 +99,9 @@ namespace UnityEngine.Rendering.HighDefinition
         // RTR
         RaytracingReflectionDirectionGeneration,
         RaytracingReflectionEvaluation,
-        RaytracingReflectionUpscaleGeneration,
-        RaytracingFilterReflection,
+        RaytracingReflectionAdjustWeight,
+        RaytracingReflectionFilter,
+        RaytracingReflectionUpscale,
         // RTAO
         RaytracingAmbientOcclusion,
         RaytracingFilterAmbientOcclusion,